change to DoubleTensor in imputation

+++ main_benchmark.py
@@ -667,7 +667,7 @@ if __name__ == "__main__":
         adj = adj.tolist()
         adjdense = sp.csr_matrix.todense(adj)
         adjsample = torch.from_numpy(adjdense)
-        adjsample = adjsample.float()
+        adjsample = adjsample.type(torch.DoubleTensor)
 
         # generate celltype regularizer from celltype
         celltypesample = generateCelltypeRegu(listResult)
